modejs: fix building for mingw32 sdk
Building hipster software like nodejs for win/mingw on a linux host is
like pure magic, but we do the attempt anyways. And we do this because
we need the nodejs-source package for the sdk, so the recipe fails if
the nodejs binary doesn't build.

Fix the build by splitting out a few libraries (c-ares, libuv),
disabling some features (perfctr, etw, dtrace), and by patching the
source in a few places. (But only patch it when building for mingw).