mongodb: fix check for unclean shutdown in init script
When using a non-default path for the mongo database files (e.g. /db),
the check for existing lock file looked at the wrong path (/data/db).
This caused the initscript to fail starting after unclean shutdowns.

Fix this by avoid expansion of $MONGOD_DB_PATH when assigning the value
for $MONGOD_LOCK_FILE, and use the former explicitly when checking the
lock file.